requirements.txt
coverage==5.0.4
django_redis==4.11.0
factory-boy==2.8.1
pre-commit==2.1.1
